backend/requirements.txt
aiogram==3.3.0
fastapi==0.110.0
uvicorn[standard]==0.27.1
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.5.3
httpx==0.27.0
phonenumbers==8.13.37
Jinja2==3.1.4
